Commonize annotation types for both box and highlight
